+++ beacon-chain/p2p/testing/mock_peermanager.go
@@ -47,8 +47,8 @@ func (m *MockPeerManager) DiscoveryAddresses() ([]multiaddr.Multiaddr, error) {
 	return m.DiscoveryAddr, nil
 }
 
-// RefreshENR .
-func (*MockPeerManager) RefreshENR() {}
+// RefreshPersistentSubnets .
+func (*MockPeerManager) RefreshPersistentSubnets() {}
 
 // FindPeersWithSubnet .
 func (*MockPeerManager) FindPeersWithSubnet(_ context.Context, _ string, _ uint64, _ int) (bool, error) {