+++ proto/prysm/v1alpha1/metadata/metadata_interfaces.go
@@ -10,6 +10,7 @@ import (
 type Metadata interface {
 	SequenceNumber() uint64
 	AttnetsBitfield() bitfield.Bitvector64
+	SyncnetsBitfield() bitfield.Bitvector4
 	InnerObject() interface{}
 	IsNil() bool
 	Copy() Metadata